[#119] Some minor refactoring to reduce amount of code, and collect the factory methods in JsonPathMatchers
