--- json-path-assert/src/main/java/com/jayway/jsonpath/matchers/IsJsonString.java
@@ -1,15 +1,0 @@
-package com.jayway.jsonpath.matchers;
-
-import com.jayway.jsonpath.ReadContext;
-import org.hamcrest.Matcher;
-
-public class IsJsonString extends IsJson<String> {
-
-    IsJsonString(Matcher<? super ReadContext> jsonMatcher) {
-        super(jsonMatcher);
-    }
-
-    public static Matcher<String> isJsonString(final Matcher<? super ReadContext> matcher) {
-        return new IsJsonString(matcher);
-    }
-}
