+++ json-path-assert/src/main/java/com/jayway/jsonpath/matchers/WithJsonPath.java
@@ -1,35 +1,20 @@
 package com.jayway.jsonpath.matchers;
 
-import com.jayway.jsonpath.*;
+import com.jayway.jsonpath.JsonPath;
+import com.jayway.jsonpath.JsonPathException;
+import com.jayway.jsonpath.PathNotFoundException;
+import com.jayway.jsonpath.ReadContext;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
-
-import static org.hamcrest.Matchers.*;
 
 public class WithJsonPath<T> extends TypeSafeMatcher<ReadContext> {
     private final JsonPath jsonPath;
     private final Matcher<T> resultMatcher;
 
-    private WithJsonPath(JsonPath jsonPath, Matcher<T> resultMatcher) {
+    public WithJsonPath(JsonPath jsonPath, Matcher<T> resultMatcher) {
         this.jsonPath = jsonPath;
         this.resultMatcher = resultMatcher;
-    }
-
-    public static Matcher<? super ReadContext> withJsonPath(String jsonPath, Predicate... filters) {
-        return withJsonPath(JsonPath.compile(jsonPath, filters));
-    }
-
-    public static Matcher<? super ReadContext> withJsonPath(JsonPath jsonPath) {
-        return withJsonPath(jsonPath, not(anyOf(nullValue(), empty())));
-    }
-
-    public static <T> Matcher<? super ReadContext> withJsonPath(String jsonPath, Matcher<T> resultMatcher) {
-        return withJsonPath(JsonPath.compile(jsonPath), resultMatcher);
-    }
-
-    public static <T> Matcher<? super ReadContext> withJsonPath(final JsonPath jsonPath, final Matcher<T> resultMatcher) {
-        return new WithJsonPath<T>(jsonPath, resultMatcher);
     }
 
     @Override
